requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-mysql
clickhouse-driver
clickhouse-connect
airflow-clickhouse-plugin
astronomer-cosmos